fix(shared): formatage unifie — preview builder = rendu composant
formatKPIValue (shared, previews des apps) et les formatters de core
(composants) divergeaient (#317) : euro -> 2 decimales (shared) vs
0 decimale (core formatCurrency) ; '%' -> suffixe texte (shared) vs
style:'percent' (core, espace insecable). La preview du builder
n'affichait pas ce que rendrait le composant.

- la famille canonique (formatValue, formatNumber, formatPercentage,
  formatCurrency, formatDecimal, formatDate) demenage dans
  @dsfr-data/shared — le contrat est celui des COMPOSANTS
- packages/core/utils/formatters.ts re-exporte (meme implementation,
  testee par identite) et garde ses helpers seuils/couleurs
- formatKPIValue : wrapper deprecie mappant unit -> format
  ('€'/'EUR' -> euro, '%' -> pourcentage, defaut -> nombre)
- politique % documentee : la valeur EST le pourcentage (5 -> '5 %')
- tolerance looksLikeNumber (detection conservatrice, numeric-auto) vs
  toNumber (parseur tolerant) documentee comme VOULUE

AC verifie : parite preview/composant sur les 4 formats KPI.
2 tests realignes (ils epinglaient la divergence).

Closes #317
